docs: create a changelog note, documenting difference between previous and new zip handout
Referring to this post from the professor on BlackBoard:

```
Thanks to some students, we have identified a bug in the evaluation function of the project. The original one was P80, which contradicts the text below it. After confirming with the companies, we have corrected the evaluation function to P20, which associates a higher weight with the overstimation.

I have updated the slides and project materials. Please download the latest ones.
```