feat: basic xgboost train & predict implementation
I asked the chatbot to set up the simplest possible working example.
This is probably as simple as it gets.

I will need to look closer at how David did the data cleaning of the
receivals, as I now simply imported receivals_cleaned, and needed to do
a magic `dt.tz_localize(None)` to allow comparison between pd.Timestamp
and DateTime object.

Currently the features are only *rm_id*, *interval_days* and *year*.
This setup does surprisingly poor, as it scores 24144, which is way
worse than 0-prediction which yields 14736.

Pain point:

Creating the train dataframe took me about 3-4 minutes. It's a
single-threaded program because of the for-loops. Should look into how
to make that section more efficient.

Generally the code needs to be a **lot more** modular. The different
steps need to be made explicit.

By steps I mean:
- data cleaning
- train dataset creation
- model creation
- model training [cross-validation, etc.]
- model prediction & submission creation