feat: add features to basic_xgboost
The new features are `month` and `day_of_week`.
The prediction dataframe needed to be updated to be compatible with the
new feature format.

It was thought that *day of week* would be relevant as there seem to be
no incoming deliveries on Saturdays and Sundays.

Function added at the end to view the top20 predicted weights across the
raw materials. This is useful for seeing if the results are *sort of
good* before sending a prediction to kaggle.

The predictions which are in the millions range are far more important
than those who are down in 100 000 (due to how the evaluation metric is
defined).

Finally, some hyperparameters for XGBoost were changed.
- n_estimators 1000 (500)
- learning_rate 0.01 (0.05)
- max_depth 20 (8)

These changes seem to be poor, as they make the model overfit too
much on the 2024 data.