pred: add bad prediction & modify basic_xgboost
This one scored 24689 on Kaggle, so quite horrible.
I tried incorporating more years (2021, 2022, 2023, 2024) and made the
max_depth smaller (back to original 8), hoping it would help mitigate overfitting.

It fitted to something which was really bad, so it seems that the
previous years contain a lot of data which aren't really good when using
XGBoost.

It feels to me that there are some assumptions the XGBoost model is
using which aren't good. Surely it will help to set the datatype of year
to categorical, as Max suggested ...
No, but at that point you totally lose the concept that 2025 follows
2024, as the categoricals are thought of as completely different
entities.