feat: add clean_materials_utils file
This is the start of a file which is going to help to the initial
cleaning of our date (before the feature engineering is done).

It currently mainly has a function to reduce the amount of rows to 1 for
each rm_id, and the row which will be kept is the one with the highest
product version.