feat: add tests for data cleaning
This ensures that the assumptions we have about the data hold.
Includes two un-implemented tests.
Currently we're only testing that the utils file is able to reduce a
dataFrame in-place such that it only contains one row for each rm_id,
where the row which is kept is the one where the product version is
highest.