fix: make validation.ipynb create 2024*0.7 prediction
You can now simply run all cells, and the test2024.csv file will be
made. That specific file is further down the line used in
`visualization/deprecated_materials.ipynb` to list the top material
predictions for this `test2024.csv` file which got a score of 9546 on
kaggle.

Previously you would need to skip one cell. That cell was simply
removed. It implemented an idea (not predict until 7 days after receival
has come) which was scrapped. (It probably turned out to be worse than
default).