test: rename test file and add new test
Test file was previously named `remove_deleted_materials.py`, but in
actuality this became more of a file for testing the general properties
of the materials.csv file.

One style change where an f-string was moved to it's own line, reducing
the line length of the code.

Added one test which looks at (product_id, product_version) combos,
showing that there are some (product_id, product_version) combos which
are listed up to 12 (!?) times.
This was (91901460.0, 5.0)
This means that the product_id is associated with 12 different rm_ids.