feat: add 0.1, 0.2 and 0.9 quantiles to DEADGE.ipynb
With the current setup, the transformer can predict using four different
quantile losses. You have 0.1, 0.5, 0.9 and 0.2

To extract the predictions you get from the 0.2 quantile, you do:
output = predictions.output.prediction[:, :, 1]
to get the 0.5 quantile.
Use 3 to get the 0.2 quantile, etc.

David is saying that 0.2 quantile is causing the model to predict 0s for
everything. This is correct. Thus, our best results come from doing 0.5
quantile and multiplying it by 0.7