refactor(preprocessing_materials): only return active_rms
It doesn't make sense to have a product_id to rm_id mapping, as one
product_id can map to several rm_ids.
This is proven in `christian/tests/test_materials_csv.py`

With this method, we go down from 203 rm_ids to something like 170.
However, David is sure that we can remove even more rm_ids from our
consideration.
We have methods like *only including the materials which were received
in 2024*. And if we do that, we end up with something like 50 rm_ids.