chore: find the code used to generate csv files in `data_cleaned/`
By using `git log -- orders_with_receivals_detailed` I was able to track
the file addition down to commit `bc736b1`, from the 19th of September.

I copied the code which there was used to generate the files.

It's a bit of a stretch to give the receivals file the name
`receivals_cleaned`, as ... you've only converted the date_arrival
string to `pd.Timestamp`.

There are 122590 rows, 57 of which are missing rm_id.
At the very least you can inspect the source code yourself now to see
how these files were generated.